+++ src/pynwb/ndx_microscopy/testing/_mock.py
@@ -489,6 +489,7 @@ def mock_MicroscopyResponseSeries(
     starting_time: Optional[float] = None,
     rate: Optional[float] = None,
     timestamps: Optional[np.ndarray] = None,
+    microscopy_series: Optional[ndx_microscopy.MicroscopyResponseSeries] = None,
 ) -> ndx_microscopy.MicroscopyResponseSeries:
     series_name = name or name_generator("MicroscopyResponseSeries")
 
@@ -525,6 +526,7 @@ def mock_MicroscopyResponseSeries(
         starting_time=series_starting_time,
         rate=series_rate,
         timestamps=series_timestamps,
+        microscopy_series=microscopy_series,
     )
 
     return microscopy_response_series