+++ CHANGELOG.md
@@ -17,6 +17,7 @@
   - Renamed `Segmentation2D` to `PlanarSegmentation`
   - Renamed `Segmentation3D` to `VolumetricSegmentation`
 - Added `get_FOV_size()` method to both `PlanarImagingSpace` and `VolumetricImagingSpace` classes for calculating Field of View size in micrometers Sub-Issue[#53](https://github.com/catalystneuro/ndx-microscopy/issues/53)
+- Added `microscopy_series` link to `MicroscopyResponseSeries` class Sub-Issue[#58](https://github.com/catalystneuro/ndx-microscopy/issues/58)
 
 # v0.2.1 (March 28, 2025)